Multiple changes
Added selected class to nav in HTML and updated CSS to indicate which
page was selected.
Added about and contact pages.
Added contact links and changed the color of the contact links, both
backgroud and hover to better match the photos and theme.
Changed the location color from yellow to white to better match the
theme.
Attempted to write out my full phone number and email in the contact
details section but they didnt fit the mobile screen so I changed them
to Keith's email and Keith's phone. Seemed to work